backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy==2.0.40
pymysql==1.1.1
python-dotenv==1.0.1
